Remove conditional button hiding logic (#508)
Previously, with the "hideDefaultButton" config option set to true, we would conditionally hide or show the button based on the Overlay's expanded/collapsed state. Removing that logic to always hide the button if "hideDefaultButton" is true.

TEST=manual

Use custom selector and set hideDefaultButton to true. Click the custom button selector and see Overlay open and close without button appearing at all.