Move offset to parent container in Overlay (#509)
The offset was on the Answers frame's iframeContainer rather than the parent container. This was previously fine because we didn't have a button iframe in addition to the other frames; but now, the translation has to happen to the parent.

TEST=manual

Test locally with a horizontal offset of -100px, see entire Overlay shift left 100px. Test with a vertical offset of -16px, see Overlay shift to be flush with the bottom of the screen.