html-asset-loader: use webpack alias and absolute path for static assets (#482)
This commit fixes the relative path calculation in the html-asset-loader.
It also adds a webpack alias to the static folder, so that the loader can
use the static asset paths regardless of the relative path of the
html file it's loading. I.e. if es/index.html and index.html
both use a static/assets/images/yext-logo.svg, they
both import them with the same require, 
`require('static/assets/images/yext-logo.svg')`,
instead of one the first being:
`require('../static/assets/images/yext-logo.svg')`,
and the second being:
`require('./static/assets/images/yext-logo.svg')`,

The isAbsoluteUrl helper had to be updated to recognize urls
that only begin with a single slash as absolute urls as well.

J=SLAP-881
T=https://yextops.zendesk.com/agent/tickets/366842
TEST=manual

test that the WHO site can specify an iconUrl that points to a static
asset

tested that a fresh site can also specify an iconUrl in a CTA
that points to a static asset, and also can directly put in
img tags with a static asset src