rename isAbsoluteUrl to isNonRelativeUrl
The hbs helper isAbsoluteUrl was also matching
on urls that only started with a single backslash,
since those don't need {{relativePath}} to be appended
to them. Technically, though, those urls are root-relative urls,
not absolute urls.

TEST=manual

test that a static asset iconUrl in the standard card cta,
and a favicon that points to a static asset, both work