Collapsible Filters: collapse filters when above results (standard/grid)
On vertical-standard/grid w/collapsible filters, when the page is
<768px wide, we get collapsible filters, and when the page is
>1200px wide we get the filters on the left sidebar. In between that,
the filters go into an intermediate state where they are above the results.
This commit replaces that intermediate state with collapsible filters.

J=SLAP-833
TEST=manual

test that on a vertical standard page with collapsible facets + sort options,
there is no more intermediate state.