Ensure that TITLE_CLICK event options contain URL
This PR ensures that the title URL is added to `titleEventOptions` for any card. The `titleEventOptions` is the payload we include with the corresponding `TITLE_CLICK` Analytics event. 

J=SLAP-920
TEST=manual

Built a local experience. Clicked on a standard card's title link and saw the URL included in the Analytics payload.