Ensure that TITLE_CLICK event options contain URL (#529)
This PR ensures that the title URL is added to `titleEventOptions` for any card. The `titleEventOptions` is the payload we include with the corresponding `TITLE_CLICK` Analytics event. The `location-standard` card and it's multilang variant were updated to call `titleUrl` simply `url`. This is in keeping with the convention in all other cards. 

J=SLAP-920
TEST=manual

Built a local experience. Clicked on the title link of various cards and saw the URL included in the Analytics payload.