Add postimport script
This is a script that will be automatically be run after jambo import.
An npm install is run after copying over the theme's package and package-lock
.json files. This is so that the script can have access to comment-json.
Otherwise, comment-json would already have to be present in node_modules,
either through a local jambo (which doesn't work if using a globally
installed jambo) or some other way.

J=SLAP-833
TEST=manual

check that all needed files are copied over, the
defaultTheme in jambo.json is updated, and layout
files are copied over and added to jambo.json
partials