html-asset-loader: trailing escapes, single quotes, relativePath in card templates
This commit updates the html-asset-loader to not match on trailing escape
characters and the double quote. The issue was that, a double quote inside
of a card template needs to be escaped two times in a webpack source. The
first time is inside the registerTemplate string, because babel will take
the template string backtik and change it into double quotes. The second
time is within the webpack source itself, where the source is put into a
(see new tests/fixtures). Pre theme 1.17, this extra escape would be removed by
the html-loader's minifyJS option. However, this option was turned off in
1.17 to increase the webpack speed, but the html-asset-loader should be
able to handle this.

Also updated the regex to match on single quotes as well as double quotes.
html-loader currently will transform all single quote html attributes,
like src='static/assets/blah.jpg', to double quotes, but it will not do
that within a card's template.hbs (because it's a js string).

also moved the regex out of the webpack config into the loader itself,
since it's more of an implementation detail. you can still override
it, if you want, but wouldn't recommend this to most hitchhikers.

finally adds a github workflow to run tests

TEST=manual, auto

test that the jinya ramen bar site can specify static assets in its template.hbs-es
again.

test that I can still specify a static asset inside component.js

test that relativePath is still handled correctly