html-asset-loader: trailing escapes, single quotes, relativePath in card templates (#535)
This commit updates the html-asset-loader to not match on trailing escape
characters. The issue was that, a double quote inside
of a card template needs to be escaped two times in a webpack source. The
first time is inside the registerTemplate string, because babel will take
the template string backtik and change it into double quotes. The second
time is within the webpack source itself, where the source is put into a
(see new tests/fixtures). Pre theme 1.17, this extra escape would be removed by
the html-loader's minifyJS option. However, this option was turned off in
1.17 to increase the webpack speed, but the html-asset-loader should be
able to handle this.

the regex was also not suited to match on {{relativePath}} prefixes in
card templates, since the relativePath in cards is computed at runtime
the regex had to also be able to match on the handlebars logic somebody
could use to specify the path.

Also updated the regex to match on single quotes as well as double quotes.
html-loader currently will transform all single quote html attributes,
like src='static/assets/blah.jpg', to double quotes, but it will not do
that within a card's template.hbs (because the template is a js string).

also moved the regex out of the webpack config into the loader itself,
since it's more of an implementation detail. you can still override
it, if you want, but wouldn't recommend this to most hitchhikers.

finally adds a github workflow to run tests

TEST=manual, auto

test that the jinya ramen bar site can specify static assets in its template.hbs-es
again (also tested adding relativePath in)

test that I can still specify a static asset inside component.js like a cta iconUrl

test favicons (tested with pages at root output directory level, and also nested in a directory)

test that both single quotes and double quotes around the static asset work