remove extra parenthesis in CssRelativeUrlTransformer (#541)
The css url transformer was adding an extra parenthesis in
inlined css when there was a relative path. This stopped
the url from 404ing, but also caused that specific css url() attribute
to be ignored (since the extra parens makes it it malformed).

T=https://hitchhikers.yext.com/community/t/answers-multi-language-repo-custom-font-not-appearing-in-secondary-locale/2103
TEST=manual, auto

test that the weight watchers site can use it's national2 font in its
en_GB locale again. see that national2.hash123.woff is requested for
in the networks tab.