product-prominentimage card: align CTAs to bottom
This change introduces additional CSS to the
product-prominentimage card so that CTAs will align
to the bottom of each card instead of appearing
immediately after the preceding content.

J=SLAP-672
TEST=manual
On a local site, used the rosetest account's products
vertical in a vertical-grid page and saw that the
cards had their CTAs aligned to the bottom.
Also tested on IE11 using browserstack to verify
UI changes looked good there.