UI Refresh: Font Updates (#548)
Update fonts for the UI Refresh

Update the font from "Open Sans" to "Source Sans Pro" and bump the sizes of every font by 2px. Add a new lighter font with a weight of 300. The old font is commented out so that the HHs can easily revert if they'd like. Make some additional updates to font sizes and font weights as well. Text that contrasts the light font weight is bumped down from semibold to normal weight. Also remove the uppercase text transform on header titles.

J=SLAP-981
TEST=visual

Generate a jambo site with these changes and visually inspect the site