UI Refresh: Update Spacing (#549)
Updating spacing to match the UI Refresh mocks

The central change is the update of the base spacing from 16px to 20px. I felt that this was the simplest way to left align all the text with 20px of spacing to match the mocks. This has the side effect of modifying the padding on all the other cards. I visually inspected the other cards to make sure the spacing update is reasonable. I didn't have mocks for all of the cards so I adjusted the padding of the other cards so that it is roughly in line with the other mocks.

Additional changes include updates to the nav, search bar, and section spacing to match the mocks.

J=SLAP-982
TEST=visual

Build a site with these spacing updates and ensure that it looks like the mocks