Testing Site (#555)
Add a testing site

Add a test site so that Percy can have something to take snapshots of. The test site can also be used for manual testing of the theme.

The testing site uses the Slapshot testing account and all the defaults of the theme. Add a setup-test-site, build-test-site, and a serve-test-site command. 

J=SLAP-834
TEST=manual

Run `npm run setup-test-site`, `npm run build-test-site` and `npm run serve-test-site` and observe the test site