reduce runtime template whitespace and remove js comments
This commit overrides the #babel helper from jambo,
and adds comments: false to it, to strip out js comments.

It also wraps all {{{read 'runtime-template' }}} usages
with {{#minifyRuntimeTemplate}}, which replaces all
instances of new line + any amount of white space with
just a single space.

TEST=manual

tested that the yanswers jp and yanswers en work with this change