Add Percy (#556)
Add Percy through PercyScript

I added a custom wait function which waits until the HTML stabilizes before returning . I found this to be the simplest way to ensure that Handlebars is done rendering before we take the screenshot.

The openStatus formatter text depends on the current time, so I added an override to the Date object constructor to return a fixed time. This ensures that percy does not create a diff based on the time that percy runs.

Due to the nature of how Percy works, the mobile screenshots of the nav bar are inaccurate. Percy takes a snapshot of the DOM, sends it to its server, and then resizes the window before taking the screenshots. Because the reshaping of the nav bar to create the dropdown menu happens in javascript, the drop down menu isn't created in the percy snapshot. Only CSS media query changes appear in the Percy snapshots. I tested enabling javascript on Percy, however that had the side effect of creating two search bars and occasionally two nav bars. Debugging that is out of scope of this item, so I opted to leave javascript disabled for now.

J=SLAP-834
TEST=visual

Run percy locally and observe the percy screenshots