Pre-load the SourceSans fonts now used by default in the Theme.
This PR adds pre-load links to fetch the new fonts used in the Theme. These
fetches are initiated before any component is actually rendered on the page.
This makes it much more likely that we will not see any "font flashing"
behavior. If a component happens to be rendered before a font request is
complete, we still have the `font-display: swap` directives. With this, the
text would be visible, albeit in a different font, until the requested font
loads. In my testing, the requests all resolved before components were
rendered.

J=SLAP-1008
TEST=manual

Tested the following:

- Saw no more "font flashing" with this change.
- On a single language site, all the fonts were loaded properly.
- On a multi-language site, the fonts were loaded properly for all localized
  pages. This includes pages that had a `relativePath` other than `.`.