Filters width fix (#575)
Decrease the breakpoint of collapsible filters

This new breakpoint is the same as the breakpoint used before the UI refresh: 1280px.

In order to be able to collapse the filters at this breakpoint, the answers container needs to shrink so that the filters don't go off the page. We can accomplish this by setting the container width to a percentage for widths larger than the cfilters breakpoint. This allows the container width to decrease smoothly to accommodate filters. At the cfilters breakpoint, the container width expands back out to full size if the space is available.

J=SLAP-999
TEST=visual

Build a site and resize a page with filters and make sure that the filters never go off the screen. Test various container widths and cfilters breakpoints and confirm that the container width decreases properly and always leaves room for the filters before hitting the cfilters breakpoint. Test in an iframe.