Add the new highlightField formatter. (#582) (#583)
This PR adds the new highlightField formatter to the Theme. This formatter is
the core of the Theme's document search support. It allows different snippets
of a string to be highlighted. Highlighted text is wrapped with '<mark>' tags.
Initially, the spec called for the tags to be configurable. However, it was
later decided that if a user wanted to change the wrapping tag, they could
just override the formatter itself.

J=SLAP-993
TEST=auto

Added unit tests for the new formatter. The back-end is not yet ready, so I
could not verify manually. Once I can test manually, I will.