Add the new `documentsearch-standard` card.
This PR adds the `documentsearch-standard` card, which is meant to handle the
new `FEATURED_SNIPPET` direct answers. I could not find any mocks for the card,
so I relied on the screenshot in the spec. This card has considerably different
styling than the existing `allfields-standard` card. Because of these things,
there will likely be some iteration with Product on the UI. I will wait to
make the multi-language variant of this card until we have buy-in on this
version.

The base `card_component` for direct answers was modified so that the snippet
could be passed to the different `dataForRender` methods.

J=SLAP-995
TEST=manual

The back-end is not yet complete, so I mocked a response with a
`FEATURED_SNIPPET` direct answer. On a simple test site, I used this mocked
response and the new card. I verified that the card's styling matched the
image in the spec. By removing pieces of the mocked response, I verified that
the card gracefully handled missing data. I ran all my tests in desktop and
mobile view.