handle translated CTA linkTypes (#585)
In the built in Calls To Action field in Knowledge Graph,
linkTypes are translated from Phone or Email to the locale
of the request. THe @yext/cta-formatter package only handles
linkTypes of either "Phone" or "Email". This caused the formatter
to fail for say a "telefono" linkType.

As a short-term solution we'll manually translate these to english.
These translations were taken from the schema for the Calls To Action
field.
Long-term, we should probably ask whoever made this field type if
its possible to not translate linkType.

J=SLAP-1029
T=https://yextops.zendesk.com/agent/tickets/379965
TEST=manual,auto

console logged the formatter's output from within a card's dataForRender().