Add the new `vertical` custom command to the Theme. (#586) (#588)
This PR adds the new `vertical` custom commad. This command creates a page for
the given vertical and sets the verticalKey and cardType placeholders in the
JSON configuration. The command takes in the following arguments (all
required): `name`, `template`, `cardName`, and `verticalKey`. A `locales`
option was not included. This is because localized sites usually use different
vertical keys and cards than their default locale counterparts.

In the `describe` method, we provide the valid options for templates. The
`page` command's logic for fetching the valid template types was reused.

J=SLAP-1041
TEST=manual, unit

Updated the `test-site` so one of the pages is built with this command. Manually tested the following:
- The output of `npx jambo vertical --help` was correct.
- The output of `npx jambo describe` was correct for all commands, including
  `vertical`.
- Running `npx jambo vertical` with the various flags set created a correctly
  configured vertical page.