Add card name validation to the `vertical` command. (#592) (#593)
Logic was added to the command to validate the provided `cardName`. If the provided card does not
exist, a `UserError` will be thrown with a helpful error message.

TEST=manual

On a test site, provided an invalid `cardName` and saw the error. The command worked as expected when I provided a valid card name.