Add the new `documentsearch-standard` card. (#584) (#597)
This PR adds the `documentsearch-standard` card, which is meant to handle the new `FEATURED_SNIPPET` direct answers. I could not find any mocks for the card, so I relied on the screenshot in the spec. This card has considerably different styling than the existing `allfields-standard` card. Because of this, there was some back and forth with UX and Product. A couple new variables were added to `answers-variables.scss`. These variables control the text highlighting for featured snippets, as well as the font weight for a direct answer card's 'View Details' text. We will have an item next Sprint to cutover the `allfields-standard` card to this new UI. 

The base `card_component` for direct answers was modified so that the snippet could be passed to the different `dataForRender` methods.

Screenshots of the new card are included below.

J=SLAP-995
TEST=manual

On a simple test site, I used the new card. I verified that the card's styling matched the image in the spec and the styling agreed upon by Product. I removed pieces of a mocked direct answer, to verify that the card gracefully handled missing data. I ran all my tests in desktop and mobile view.