Update the default `allfields-standard` card styling. (#603) (#604)
This PR updates the styling for the `allfields-standard` card. Specifically,
select styling from the `documentsearch-standard` card was copied over. Note
that the header of the card was left unchanged, per Product's request.

J=SLAP-1064
TEST=manual

Created a local test site with the updated card styling. Verified the UI and UX
with Rose. Confirmed both mobile and desktop styling.