interactive-map: Add Consulting Locator code
All of the code files here include code from Consulting.
It includes some modifications used to allow for functionality
we have on our interactive map component.

The idea in the future is to have this code live in their
node module and to import this code directly.

J=SLAP-1081
TEST=manual

This code (entry from Maps/Map.js) is used in the interactive-map.
Make sure that the Map from Map.js is correctly allows a map
and interactive behavior in the component.