interactive-map: Add InteractiveMap component and page template (#606)
This PR copies the vertical-map page template and then further edits
it to be the new InteractiveMap page template. This template features
a new InteractiveMap component that acts on the entire page DOM to
fix interactions between different pieces of the page.

The goal of this component is to allow a place to pass in configuration
via the JSON and have that manifest in different behaviors of the map
and results card. For example, we want people to be able to modify
pin colors and whether or not the Search This Area happens by default.
This component reads configuration and then assigns behavior
accordingly on the page.

The map interacts with Consulting code to take advantage of already
tested behavior like clustering and pin states.

The InteractiveMap does not have its own hbs template, rather it
relies on certain DOM elements existing on the page already (in the
page.html.hbs file). This is because we don't have built-in top-level
component support in the SDK and theme. Components like LocationBias
don't have built in configuration for waterfall behavior (there is the
onMount but that relies on implementation details in the SDK). In the theme,
this would change the way HHs would be able to interface with adding,
moving, and removing components. To remove a component, they would
need to either change some InteractiveMap component template (outside
of the normal placement of templates) or they would need to shadow the
component js and remove the correct scripts.

Note: While we use the page.html.hbs for most of the component creation
and configuration, some live in this top-level component. The two special
components are
(1) The AlternativeVerticals component: This is because the AlternativeVerticals
requires a parent component to pass it data. It does not have a listener for
vertical results
(2) The new Map component: This is because we must pass in listeners that
interact with elements outside of the map (for instance, we must tell the
results list to scroll when a pin is clicked). These functions live inside of the
InteractiveMap, and is not easily available to the standalone Map component.

Note: We keep the configuration of pin images in two files (below). This
is meant to be a file that is shadowed by HHs often, and given the flexibility
to do things like change pin images based on profile information. Ideally we
would make these files as user-friendly as possible.
ClusterPinImages.js
PinImages.js

J=SLAP-1081
TEST=manual

See an interactive map vertical in a Jambo-built theme site with the
desired interactivity. For example, clicking a card applies selected styling
for a pin. Also clicking on the Map toggle button on mobile shows a mobile
map.

Using the test prod account with the locations vertical. There is other
setup in getting the page working with webpack-config and core.hbs
imports that will be seen in the next PR.