interactive-map: Hook up the component to the theme (#610)
This PR makes changes to the existing theme parts to integrate the
InteractiveMap component into theme. This includes:

SCSS changes
making changes to the location-standard card to handle the new
InteractiveMap card layout
Adding a new webpack entry for the new theme components
Importing that entry in the core.hbs
Creating the addComponent call for the InteractiveMap
With the addition of this PR, one should be able to create an
interactive map vertical with the code in the feature.

J=SLAP-1081
TEST=manual

See an interactive map vertical in a Jambo-built theme site with the
desired interactivity. For example, clicking a card applies selected styling
for a pin. Also clicking on the Map toggle button on mobile shows a mobile
map.

Using the test prod account with the locations vertical.