Add mapbox support for the map event listeners
All event listeners for the Search This Area event listeners
were specific to Google. Add these event handlers to the general
ProviderMap class to support this in multiple providers (Mapbox
and Google in our case).

We make a few changes to the event listeners:
* Remove the map parameter. This is so we do not pass implementation
details to the listener functions. Instead we pass in zoomTrigger
directly for the zoom listener.
* zoomChange -> zoomEnd to mark that it is when the zoom is over

Also: fix pin clustering so it could be tested

J=SLAP-1056
TEST=manual

Test that a
* Click on cluster click will call search this area
* Dragging the map will trigger a search this area
* Zooming the map will trigger a search this area

On both Mapbox and Google

Address zoomEndListener

Debounce searches on map movement (#611)

Prevent searching twice when the map has only been moved a little

The debouncer takes into account various zoom levels and only allows a search to be ran again once it is outside the distance threshold. The threshold is based on a percentage of the map width. This PR configures the debouncer to only allow a new search if the map has moved 20% of the map width away from the previous search.

J=SLAP-1070
TEST=manual

Test the dragging the map around at various zoom levels and confirm that the map needs to move a certain amount before a new search is ran. Confirm the behavior is the same regardless of the zoom level. Test that a new search is ran when the zoom buttons are pressed.

Changes to fit search debouncer with new listeners