Debounce searches on map movement (#611)
Prevent searching twice when the map has only been moved a little

The debouncer takes into account various zoom levels and only allows a search to be ran again once it is outside the distance threshold. The threshold is based on a percentage of the map width. This PR configures the debouncer to only allow a new search if the map has moved 20% of the map width away from the previous search.

J=SLAP-1070
TEST=manual

Test the dragging the map around at various zoom levels and confirm that the map needs to move a certain amount before a new search is ran. Confirm the behavior is the same regardless of the zoom level. Test that a new search is ran when the zoom buttons are pressed.