Add mapbox support for the map event listeners (#612)
All event listeners for the Search This Area event listeners
were specific to Google. Add these event handlers to the general
ProviderMap class to support this in multiple providers (Mapbox
and Google in our case).

We make a few changes to the event listeners:

* Remove the map parameter. This is so we do not pass implementation
details to the listener functions. Instead we pass in zoomTrigger
directly for the zoom listener.
* zoomChange -> zoomEnd to mark that it is when the zoom is over
* Restructure the zoomTrigger logic to better know when a zoom
has ended and what triggered it.
* Rebase with the new debounced search

Also: fix pin clustering so it could be tested

J=SLAP-1056
TEST=manual

Test that a

Click on cluster click will call search this area
Dragging the map will trigger a search this area
Zooming the map will trigger a search this area
On both Mapbox and Google