interactive-map: Update InteractiveMap to use the newest SDK
Use the latest develop branch to take advantage of the new storage in
the SDK. Did not see many regressions as a result of this change. Copied
over changes for the collapsible filters from #607.

Note: we will keep the feature branch on develop of the SDK until we
merge it into the develop branch of the theme.

Note: there is no multilang bundle for canary/latest. To test multilang,
you should use a locally-run version of the SDK on develop.

J=SLAP-1103
TEST=manual

Thankfully, after a smoke test, no regressions were found.
Tested interactions like:

* Clicking a pin (on mobile, on desktop), applied the correct select
state, in mobile created a detail card, in desktop scrolled to the
correct card in the result list.
* Making a new search updates the map with the new results
* Debouncing only works when you've moved from the original search center enough