Only load the locator bundle on interactive map pages (#621)
Only load the locator bundle on interactive map pages

Previously the locator-components.js bundle was being loaded on every page template. Update it so that it is only being loaded on the vertical-interactive-map page

J=SLAP-1096
TEST=manual

Build a site and confirm that the interactive map still works. Build a location-standard page and confirm that the locator-components.js bundle is no longer being loaded