interactive-map: Automatically encode custom pin SVGs (#618)
This is mostly adding back previous behavior. Previously, we would add
this utf-8 encoding to the pins passed into the Mapbox/Google
implementations. Add this back in such a way that a HH can still paste
an SVG straight into the PinImages/ClusterPinImages functions without
having to remember to use this.

J=SLAP-1112
TEST=manual

Test on both Google, Mapbox
* You can see default pins
* You can see default cluster pins
* You can see their hovered, selected variants