interactive-map: Use centralized StorageKeys class (#617)
* interactive-map: Use centralized StorageKeys class

Instead of having the SDK storage keys hardcoded in multiple files, we
centralize it in one theme storage keys file. This file can be used to
store all storage keys used in the theme. This includes native SDK keys
because those are not directly exposed by the library.

Note: I took this opportunity to (in a second commit) standardize the
key names and values.

J=SLAP-1113
TEST=manual

Test that global storage interactions still work using the new
StorageKeys class.

* Clicking a card will highlight the corresponding pin on the map
* Clicking a pin will scroll to the corresponding card on the results
  list
* Search this area is still debounced when moving the cursor a tiny bit
* On a new vertical results search, the map updates