Add the 4 Column universal search layout. (#622)
* Add the 4 Column universal search layout.

This PR adds a new universal section template: grid-four-columns. This
template lays out the results in a four colum grid (as the name suggests).
The layout works best with the product type cards and, as such, has only
been tested with them. We already had an existing Hitchhiker-4-columns CSS
class that can be used in a vertical-grid template.

J=SLAP-989
TEST=manual

Built a sample site and used the new section template for the products
vertical. Also made sure that the product vertical page was using the
Hitchhikers-4-columns class. Verified things looked correct on universal and
vertical. Screenshots will be attached.

* Remove newline.