interactive-map: Fix View Results button styling on tablet
Previously, the View Results button for Collapsible Filters was not
hanging at the bottom of the screen on tablet breakpoint. We fix this
specifically for the Interactive Map page type, where tablet breakpoint
should match the View Results button behavior of mobile.

J=SLAP-1058
TEST=manual

Check that the view results button on mobile, tablet, and desktop have
the correct behavior (desktop, position absolute to the content
wrapper).