WCAG Navigation (#625)
Update the locator so that it is aria navigable with a screen reader or a keyboard

- Add focus listeners to any location card <a> element to trigger the pinFocus. I renamed pinClicked and pinClickListener to pinFocus and pinFocusListener because pinFocus better describes both a click and focus state.
- Make the scroll to pin feature work when tabbing through the pins
- Move the stickyBottom div so that cards are never hidden behind it
- Add focus states to the `searchThisAreaToggle` and the update location button.
- Add aria landmarks to the Map, Main location search section, and the Map controls section
- Move up the searchThisArea div so that you can tab into it right before going onto the map

Note: the `searchThisAreaToggle` focus doesn't work in IE11 because focus-within is not supported in IE11. Getting this to work in IE11 would likely require a re-work of the search this area checkbox, so this maybe makes more sense to tackle in QA

J=SLAP-1059
TEST=manual

In google chrome and safari, tab through the cards and observe the card being highlighted and observe the corresponding pin show focus. This doesn't work in firefox because firefox doesn't make <a> elements tabable by default. Since this is a larger issue with the theme, I did not address that in this PR. The solution likely involves adding tabindex="0" to those elements

Test that tabbing through the map pins causes the scroll to pin behavior.

Tab over the `searchThisAreaToggle` and the the update location button and view the focus state

Use the screen reader and the firefox landmarks extension and see the landmarks on the page. The role="region" is newer so sometimes it is only visible to newer accessibility tools.